CRUD METAS

+++ database/migrations/2025_06_29_032412_create_meta_plan_nacionals_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->string('nombre');
             $table->string('descripcion');
-            $table->double('porcentajeAlineacion');
+            $table->decimal('porcentajeAlineacion');
         });
     }
